# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Quick Relay VPS API",
    "version": "vps-1",
    "description": "Existing Go API after repository cleanup. URLs include /api/v1; the server base is the origin."
  },
  "servers": [
    {
      "url": "https://relay.example.com"
    }
  ],
  "security": [
    {
      "deviceBearer": []
    }
  ],
  "paths": {
    "/api/v1/pair/complete": {
      "post": {
        "summary": "Complete CLI-issued pairing",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "ok",
                    "device_access_token",
                    "expires_at"
                  ],
                  "properties": {
                    "ok": {
                      "const": true
                    },
                    "device_access_token": {
                      "type": "string"
                    },
                    "expires_at": {
                      "type": "null"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "500": {
            "description": "Internal error"
          },
          "401": {
            "description": "Invalid, expired or exhausted pairing code"
          },
          "429": {
            "description": "Rate limited; Retry-After: 60"
          }
        },
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": [
                  "pairing_code",
                  "installation_id"
                ],
                "properties": {
                  "pairing_code": {
                    "type": "string"
                  },
                  "installation_id": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/devices/register": {
      "post": {
        "summary": "Register or rotate this device APNs token",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "ok",
                    "device"
                  ],
                  "properties": {
                    "ok": {
                      "const": true
                    },
                    "device": {
                      "$ref": "#/components/schemas/Device"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "500": {
            "description": "Internal error"
          },
          "401": {
            "description": "Bearer authentication required"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": [
                  "installation_id",
                  "device_token",
                  "environment"
                ],
                "properties": {
                  "installation_id": {
                    "type": "string"
                  },
                  "device_token": {
                    "type": "string"
                  },
                  "environment": {
                    "enum": [
                      "development",
                      "production",
                      "sandbox"
                    ]
                  },
                  "app_version": {
                    "type": "string"
                  },
                  "os_version": {
                    "type": "string"
                  },
                  "device_name": {
                    "type": "string"
                  },
                  "preferences": {
                    "anyOf": [
                      {
                        "$ref": "#/components/schemas/Preferences"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/devices/me": {
      "get": {
        "summary": "Read own device without tokens",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "ok",
                    "device"
                  ],
                  "properties": {
                    "ok": {
                      "const": true
                    },
                    "device": {
                      "$ref": "#/components/schemas/Device"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "500": {
            "description": "Internal error"
          },
          "401": {
            "description": "Bearer authentication required"
          }
        }
      },
      "put": {
        "summary": "Register or rotate own APNs token",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "ok",
                    "device"
                  ],
                  "properties": {
                    "ok": {
                      "const": true
                    },
                    "device": {
                      "$ref": "#/components/schemas/Device"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "500": {
            "description": "Internal error"
          },
          "401": {
            "description": "Bearer authentication required"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": [
                  "installation_id",
                  "device_token",
                  "environment"
                ],
                "properties": {
                  "installation_id": {
                    "type": "string"
                  },
                  "device_token": {
                    "type": "string"
                  },
                  "environment": {
                    "enum": [
                      "development",
                      "production",
                      "sandbox"
                    ]
                  },
                  "app_version": {
                    "type": "string"
                  },
                  "os_version": {
                    "type": "string"
                  },
                  "device_name": {
                    "type": "string"
                  },
                  "preferences": {
                    "anyOf": [
                      {
                        "$ref": "#/components/schemas/Preferences"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Revoke own API credential and push registration",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "ok"
                  ],
                  "properties": {
                    "ok": {
                      "const": true
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "500": {
            "description": "Internal error"
          },
          "401": {
            "description": "Bearer authentication required"
          }
        }
      }
    },
    "/api/v1/devices/me/preferences": {
      "patch": {
        "summary": "Update selected preference fields",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "ok",
                    "preferences"
                  ],
                  "properties": {
                    "ok": {
                      "const": true
                    },
                    "preferences": {
                      "$ref": "#/components/schemas/Preferences"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "500": {
            "description": "Internal error"
          },
          "401": {
            "description": "Bearer authentication required"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": [],
                "properties": {
                  "notifications_enabled": {
                    "type": "boolean"
                  },
                  "time_sensitive_enabled": {
                    "type": "boolean"
                  },
                  "custom_sound_enabled": {
                    "type": "boolean"
                  },
                  "event_types": {
                    "type": [
                      "array",
                      "null"
                    ],
                    "maxItems": 20,
                    "items": {
                      "enum": [
                        "eew_forecast",
                        "eew_warning",
                        "eew_cancel",
                        "earthquake_info",
                        "earthquake_update",
                        "system_test",
                        "tsunami_warning",
                        "tsunami_info",
                        "nankai_info",
                        "seismic_advisory",
                        "earthquake_data"
                      ]
                    }
                  },
                  "earthquake_regions": {
                    "type": "array",
                    "maxItems": 20,
                    "uniqueItems": true,
                    "items": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 60
                    }
                  },
                  "tsunami_regions": {
                    "type": "array",
                    "maxItems": 20,
                    "uniqueItems": true,
                    "items": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 60
                    }
                  },
                  "minimum_intensity": {
                    "enum": [
                      "",
                      "1",
                      "2",
                      "3",
                      "4",
                      "5-",
                      "5+",
                      "6-",
                      "6+",
                      "7"
                    ]
                  },
                  "live_activities_enabled": {
                    "type": "boolean",
                    "default": false
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/devices/me/notification-tests": {
      "post": {
        "summary": "Send a labelled diagnostic to the requesting installation only",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": [
                  "request_id",
                  "style"
                ],
                "properties": {
                  "request_id": {
                    "type": "string",
                    "pattern": "^[A-Za-z0-9_-]{1,128}$"
                  },
                  "style": {
                    "enum": [
                      "normal",
                      "warning"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "ok",
                    "test",
                    "cooldown_seconds"
                  ],
                  "properties": {
                    "ok": {
                      "const": true
                    },
                    "test": {
                      "$ref": "#/components/schemas/NotificationTest"
                    },
                    "cooldown_seconds": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Installation authentication required"
          },
          "409": {
            "description": "Notifications or installation disabled"
          },
          "429": {
            "description": "One new request per 60 seconds"
          },
          "503": {
            "description": "APNs unavailable"
          }
        }
      }
    },
    "/api/v1/devices/me/notification-tests/{test}": {
      "get": {
        "summary": "Read this installation's diagnostic",
        "parameters": [
          {
            "name": "test",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{1,128}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "ok",
                    "test",
                    "cooldown_seconds"
                  ],
                  "properties": {
                    "ok": {
                      "const": true
                    },
                    "test": {
                      "$ref": "#/components/schemas/NotificationTest"
                    },
                    "cooldown_seconds": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "404": {
            "description": "Unknown or not owned"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "deviceBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "Preferences": {
        "$ref": "./api.schema.json#/$defs/Preferences"
      },
      "Device": {
        "$ref": "./api.schema.json#/$defs/Device"
      },
...
      "NotificationTest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "style",
          "status",
          "requested_at"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "style": {
            "enum": [
              "normal",
              "warning"
            ]
          },
          "status": {
            "enum": [
              "pending",
              "apns_accepted",
              "rejected",
              "result_unknown"
            ]
          },
          "requested_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    }
  }
}